+++ autotraders/ship/fuel.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class Fuel(BaseModel):
     current: int
-    total: int
+    capacity: int
 
     def __str__(self):
         return str(self.current) + "/" + str(self.total)