+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 - BREAKING: Added and enabled http2
 - BREAKING: Removed `navigate_async`, (it was deprecated)
 - BREAKING: Removed `get_session`, you can instantiate the client for greater control.
+- BREAKING: Renamed `Fuel.total` to `Fuel.capacity`
+- BREAKING: `status.leaderboards` is no longer a `list` it is now a `dict`
+- BREAKING: `status.next_reset` and `status.reset_frequency` have been moved to `status.server_reset`
+and renamed to `next` and `frequency`
+- BREAKING: `ship.reactor.cooldown` has been completely removed
 - UPDATED: Changed `attrs` to `pydantic`.
 
 # v1.9.0